Move classes into topo package
