+++ server/sink/src/main/java/org/bithon/server/sink/metrics/topo/TopoTransformers.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.sink.metrics.transformer;
+package org.bithon.server.sink.metrics.topo;
 
 import org.bithon.server.storage.meta.IMetaStorage;
 import org.springframework.stereotype.Component;
