+++ server/web-app/src/main/resources/dashboard/mongodb-metrics.json
@@ -11,7 +11,7 @@
       ],
       "metrics": [
         {
-          "name": "callCount"
+          "name": "totalCount"
         },
         {
           "name": "tps",
